implement google sign in
I added google sign in to app and user can sign in by google account. If state of account is correctly and user not exist it will create account in firebase. When user exist and didn't sign out before at the next time open app user will be sign automatically to app.