transfer repeat section form firebase to room
I changed the data retrieval model for the repeat section. from now on, the room database will be responsible for the replay section, I moved it from firebase to speed up the application. when you turn on the application, it checks the flashcards that have been learned by the user in the lesson section and if they have reached the revision date, it puts them in the revision section, but the section is limited, at the moment it is 30 flashcards. In the future, this will be expanded to allow the user to select the number of flashcards.