fun changed from firebase to room
I made a lot of changes from the firebase database to room. The first one is adding a box with flashcards from the public view to a private user and all the functions related to it, i.e. starting the learning section immediately after adding the public box to the private one.